[FEAT/#3] FSD 구조 도입 및 Biome, NativeWind 환경 설정
