+++ .github/labeler.yml
@@ -1,0 +1,74 @@
+# labeler "full" schema
+
+# enable labeler on issues, prs, or both.
+enable:
+  issues: true
+  prs: true
+# comments object allows you to specify a different message for issues and prs
+
+comments:
+  issues: |
+    Thanks for opening this issue!
+    I have applied any labels matching special text in your title and description.
+
+    Please review the labels and make any necessary changes.
+  prs: |
+    Thanks for the contribution!
+    I have applied any labels matching special text in your title and description.
+
+    Please review the labels and make any necessary changes.
+
+# Labels is an object where:
+# - keys are labels
+# - values are objects of { include: [ pattern ], exclude: [ pattern ] }
+#    - pattern must be a valid regex, and is applied globally to
+#      title + description of issues and/or prs (see enabled config above)
+#    - 'include' patterns will associate a label if any of these patterns match
+#    - 'exclude' patterns will ignore this label if any of these patterns match
+#    - 'branches' is an optional array of branch names (or patterns) to limit labeling according to PR target branch
+labels:
+  'bug':
+    include:
+      - '(?i)bug'
+      - '(?i)fix'
+    exclude: []
+  'feature':
+    include:
+      - '(?i)feat'
+    exclude: []
+  'mod':
+    include:
+      - '(?i)mod'
+    exclude: []
+  'chore':
+    include:
+      - '(?i)chore'
+    exclude: []
+  'del':
+    include:
+      - '(?i)del'
+    exclude: []
+  'merge':
+    include:
+      - '(?i)merge'
+    exclude: []
+  'move':
+    include:
+      - '(?i)move'
+    exclude: []
+  'rename':
+    include:
+      - '(?i)rename'
+    exclude: []
+  'refactor':
+    include:
+      - '(?i)refactor'
+    exclude: []
+  'docs':
+    include:
+      - '(?i)docs'
+    exclude: []
+  'deploy':
+    include:
+      - '(?i)deploy'
+    exclude: []
