+++ .github/pull_request_template.md
@@ -1,0 +1,18 @@
+## 📝 요약 
+- 
+
+## ⚙️ 작업 내용 
+- 
+- 
+
+## 🔗 관련 이슈 
+- Closes #
+
+## ✅ 체크리스트 
+- [ ] 코딩 컨벤션(Biome/Lint)을 준수하였습니다.
+- [ ] 모든 타입 에러를 해결하였습니다. (Typecheck)
+- [ ] 변경 사항에 대한 테스트를 마쳤습니다.
+- [ ] 불필요한 로그(console.log)를 제거하였습니다.
+
+## 💬 리뷰어에게 
+-
