+++ src/app/(tabs)/_layout.tsx
@@ -1,0 +1,15 @@
+import { Tabs } from "expo-router";
+
+export default function TabLayout() {
+	return (
+		<Tabs>
+			<Tabs.Screen
+				name="index"
+				options={{
+					title: "Home",
+					tabBarIcon: () => null,
+				}}
+			/>
+		</Tabs>
+	);
+}
