+++ src/app/(tabs)/index.tsx
@@ -1,0 +1,10 @@
+import { View, Text } from "react-native";
+
+export default function HomeScreen() {
+	return (
+		<View className="flex-1 items-center justify-center p-5">
+			<Text className="text-[32px] font-bold mb-2.5">Welcome</Text>
+			<Text>앱이 정상적으로 실행되었습니다112.</Text>
+		</View>
+	);
+}
