+++ src/app/_layout.tsx
@@ -1,0 +1,18 @@
+import { Stack } from "expo-router";
+import { StatusBar } from "expo-status-bar";
+import "../shared/styles/global.css";
+
+export const unstable_settings = {
+	anchor: "(tabs)",
+};
+
+export default function RootLayout() {
+	return (
+		<>
+			<Stack>
+				<Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+			</Stack>
+			<StatusBar style="auto" />
+		</>
+	);
+}
