+++ src/app/index.tsx
@@ -1,0 +1,5 @@
+import { Redirect } from "expo-router";
+
+export default function Index() {
+	return <Redirect href="/(tabs)" />;
+}
