+++ src/shared/assets/images/android-icon-monochrome.png
