+++ src/shared/assets/images/favicon.png
