+++ src/shared/assets/images/partial-react-logo.png
