+++ src/shared/assets/images/react-logo.png
