+++ src/shared/assets/images/react-logo@2x.png
