+++ src/shared/assets/images/splash-icon.png
