+++ tailwind.config.js
@@ -1,0 +1,11 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+	content: [
+		"./src/**/*.{js,jsx,ts,tsx}",
+	],
+	presets: [require("nativewind/preset")],
+	theme: {
+		extend: {},
+	},
+	plugins: [],
+};
